change actions menu to tabManager
